feat and refactor: splitting the code into different files and creating new get and insert functions
-trying to make the project more organized and solid.
