+++ moneymind/src/mysql_initializer.py
@@ -1,16 +1,14 @@
 import mysql.connector
 import os
 from dotenv import load_dotenv
-from time import sleep
 
 load_dotenv()
 
 
 def establishing_mysql_connection():
     try:
-        sleep(5)
         connection = mysql.connector.connect(
-            host=os.environ.get('MYSQL_HOST'),
+            # host=os.environ.get('MYSQL_HOST'),
             user=os.environ.get('MYSQL_USER'),
             password=os.environ.get('MYSQL_PASSWORD'),
             database=os.environ.get('MYSQL_DATABASE'),
@@ -25,6 +23,12 @@ def establishing_mysql_connection():
         )
         cursor.execute(
             "CREATE TABLE IF NOT EXISTS lista_tarifas_valores (id INT AUTO_INCREMENT PRIMARY KEY, cnpj VARCHAR(50) NOT NULL, razao_social VARCHAR(256) NOT NULL, valor_maximo VARCHAR(256) NOT NULL, periodicidade VARCHAR(256) NOT NULL, servico VARCHAR(50), grupo VARCHAR(50))"
+        )
+        cursor.execute(
+            "CREATE TABLE IF NOT EXISTS instituicoes (id INT AUTO_INCREMENT PRIMARY KEY, cnpj VARCHAR(50) NOT NULL, nome VARCHAR(256) NOT NULL)"
+        )
+        cursor.execute(
+            "CREATE TABLE IF NOT EXISTS lista_tarifas_instituicoes (id INT AUTO_INCREMENT PRIMARY KEY, codigo_servico VARCHAR(50) NOT NULL, servico VARCHAR(256) NOT NULL, unidade VARCHAR(256) NOT NULL, data_vigencia VARCHAR(256) NOT NULL, valor_maximo FLOAT NOT NULL, tipo_valor VARCHAR(50),periodicidade VARCHAR(50),cnpj VARCHAR(50),pessoa VARCHAR(50))"
         )
         connection.commit()
         return connection
